0.5.6.0
Implemented ship's movement and braking stiffness
Implemented movement direction rebound for ship's movement when near the viewport
Implemented bounds displacement config
Implemented movement stiffness config
Implemented movement finate state machines and enemy movement
Implemented key validation in saving system

Spawners behaviour improved
AuxMath and AuxAsync API expanded
Saving system behaviour improved

Bug fixed: in ship movement max velocity and max rotation always set to its highest possible values
Bug fixed: game controls transmitter might not emit current mouse world position correctly

+++ Assets/Project/Scripts/Main/Audio/Audio player/AudioPlayer.cs
@@ -216,7 +216,7 @@ namespace SpaceAce.Main.Audio
             cache.AudioSource.bypassEffects = false;
             cache.AudioSource.bypassListenerEffects = false;
             cache.AudioSource.bypassReverbZones = false;
-            cache.AudioSource.priority = (byte)properties.Priority;
+            cache.AudioSource.priority = (int)properties.Priority;
             cache.AudioSource.volume = properties.Volume;
             cache.AudioSource.spatialBlend = properties.SpatialBlend;
             cache.AudioSource.pitch = properties.Pitch;
